Build and test a Docker image of VirtualPDU
To configure the service, some environmental variables must be
set.  It is possible to pass a complete configuration to simulate
multiple units or only the settings to override if there is only one
unit to simulate.

Full configuration
------------------

In this mode, the `config` environment variable must be set with the
content of the configuration file that VirtualPDU expects.

**Example from docker-compose.yml:**

    services:
      virtualpdu:
        image: "internap/virtualpdu:latest"
        ports:
         - '9997:9997/udp'
        environment:
          config: |
            [global]
            libvirt_uri=test:///default
            outlet_default_state=ON

            [my_second_pdu]
            listen_address=0.0.0.0
            listen_port=9997
            community=public
            ports=2:test

Single PDU configuration
------------------------

In this mode, only the configuration that needs to be changed has to be
declared as an environment variable.

**Example from docker-compose.yml:**

    services:
      virtualpdu:
        image: "internap/virtualpdu:latest"
        ports:
         - '9997:9997/udp'
        environment:
          libvirt_uri: test:///default
          outlet_default_state: OFF
          listen_address: 0.0.0.0
          listen_port: 9997
          community: public
          ports: 2:test